chore(nx): cleanup storybook schematics
bump storybook versions to 5.2.5
remove @types/storybook__addon-knobs

chore(nx): remove css-loader package dependency
