+++ e2e/storybook.test.ts
@@ -1,13 +1,12 @@
 import {
-  createTestUILib,
   forEachCli,
   runCLI,
   supportUi,
-  newProject,
-  runYarnInstall,
   uniq,
-  ensureProject
+  ensureProject,
+  tmpProjPath
 } from './utils';
+import { writeFileSync } from 'fs';
 
 forEachCli(() => {
   describe('Storybook schematics', () => {
@@ -34,7 +33,6 @@ forEachCli(() => {
           runCLI(
             `generate @nrwl/storybook:configuration ${mylib2} --configureCypress --generateStories --generateCypressSpecs --no-interactive`
           );
-          runYarnInstall();
 
           expect(
             runCLI(`run ${mylib}-e2e:e2e --configuration=headless --no-watch`)
@@ -44,3 +42,47 @@ forEachCli(() => {
     }
   });
 });
+
+export function createTestUILib(libName: string): void {
+  runCLI(`g @nrwl/angular:library ${libName} --no-interactive`);
+  runCLI(
+    `g @schematics/angular:component test-button --project=${libName} --no-interactive`
+  );
+
+  writeFileSync(
+    tmpProjPath(`libs/${libName}/src/lib/test-button/test-button.component.ts`),
+    `
+import { Component, OnInit, Input } from '@angular/core';
+
+export type ButtonStyle = 'default' | 'primary' | 'accent';
+
+@Component({
+  selector: 'proj-test-button',
+  templateUrl: './test-button.component.html',
+  styleUrls: ['./test-button.component.css']
+})
+export class TestButtonComponent implements OnInit {
+  @Input('buttonType') type = 'button';
+  @Input() style: ButtonStyle = 'default';
+  @Input() age: number;
+  @Input() isOn = false;
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
+      `
+  );
+
+  writeFileSync(
+    tmpProjPath(
+      `libs/${libName}/src/lib/test-button/test-button.component.html`
+    ),
+    `<button [attr.type]="type" [ngClass]="style"></button>`
+  );
+  runCLI(
+    `g @schematics/angular:component test-other --project=${libName} --no-interactive`
+  );
+}
