+++ e2e/utils.ts
@@ -143,50 +143,6 @@ exports.default = default_1;`
   execSync(`cp -a ${tmpBackupProjPath()} ${tmpProjPath()}`);
 }
 
-export function createTestUILib(libName: string): void {
-  runCLI(`g @nrwl/angular:library ${libName} --no-interactive`);
-  runCLI(
-    `g @schematics/angular:component test-button --project=${libName} --no-interactive`
-  );
-
-  writeFileSync(
-    tmpProjPath(`libs/${libName}/src/lib/test-button/test-button.component.ts`),
-    `
-import { Component, OnInit, Input } from '@angular/core';
-
-export type ButtonStyle = 'default' | 'primary' | 'accent';
-
-@Component({
-  selector: 'proj-test-button',
-  templateUrl: './test-button.component.html',
-  styleUrls: ['./test-button.component.css']
-})
-export class TestButtonComponent implements OnInit {
-  @Input('buttonType') type = 'button';
-  @Input() style: ButtonStyle = 'default';
-  @Input() age: number;
-  @Input() isOn = false;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
-      `
-  );
-
-  writeFileSync(
-    tmpProjPath(
-      `libs/${libName}/src/lib/test-button/test-button.component.html`
-    ),
-    `<button [attr.type]="type" [ngClass]="style"></button>`
-  );
-  runCLI(
-    `g @schematics/angular:component test-other --project=${libName} --no-interactive`
-  );
-}
-
 export function ensureProject(): void {
   if (!directoryExists(tmpProjPath())) {
     newProject();
@@ -266,7 +222,9 @@ export function copyMissingPackages(): void {
 
     'next',
     'next-server',
-    'document-register-element'
+    'document-register-element',
+
+    '@angular/forms'
   ];
   modulesToCopy.forEach(m => copyNodeModule(m));
   updateFile(
