+++ package.json
@@ -36,6 +36,7 @@
     "@angular/compiler": "^8.2.0",
     "@angular/compiler-cli": "^8.2.0",
     "@angular/core": "^8.2.0",
+    "@angular/forms": "^8.2.0",
     "@angular/platform-browser": "^8.2.0",
     "@angular/platform-browser-dynamic": "^8.2.0",
     "@angular/router": "^8.2.0",
@@ -61,7 +62,9 @@
     "@ngrx/store-devtools": "8.3.0",
     "@schematics/angular": "8.3.3",
     "@testing-library/react": "9.3.0",
-    "@storybook/core": "^5.2.0",
+    "@storybook/core": "^5.2.5",
+    "@storybook/angular": "^5.2.5",
+    "@storybook/addon-knobs": "^5.2.5",
     "@types/express": "4.17.0",
     "@types/fast-levenshtein": "^0.0.1",
     "@types/jasmine": "~2.8.6",
