+++ packages/angular/src/utils/testing.ts
@@ -166,59 +166,3 @@ export function createLib(tree: Tree, libName: string): Tree {
   );
   return tree;
 }
-
-export async function createTestUILib(libName: string): Promise<Tree> {
-  let appTree = Tree.empty();
-  appTree = createEmptyWorkspace(appTree);
-  appTree = await callRule(
-    externalSchematic('@nrwl/angular', 'library', {
-      name: libName
-    }),
-    appTree
-  );
-  appTree = await callRule(
-    externalSchematic('@schematics/angular', 'component', {
-      name: 'test-button',
-      project: libName
-    }),
-    appTree
-  );
-  appTree.overwrite(
-    `libs/${libName}/src/lib/test-button/test-button.component.ts`,
-    `
-import { Component, OnInit, Input } from '@angular/core';
-
-export type ButtonStyle = 'default' | 'primary' | 'accent';
-
-@Component({
-  selector: 'proj-test-button',
-  templateUrl: './test-button.component.html',
-  styleUrls: ['./test-button.component.css']
-})
-export class TestButtonComponent implements OnInit {
-  @Input('buttonType') type = 'button';
-  @Input() style: ButtonStyle = 'default';
-  @Input() age: number;
-  @Input() isOn = false;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
-`
-  );
-  appTree.overwrite(
-    `libs/${libName}/src/lib/test-button/test-button.component.html`,
-    `<button [attr.type]="type" [ngClass]="style"></button>`
-  );
-  appTree = await callRule(
-    externalSchematic('@schematics/angular', 'component', {
-      name: 'test-other',
-      project: libName
-    }),
-    appTree
-  );
-  return appTree;
-}
