+++ packages/storybook/src/utils/versions.ts
@@ -1,5 +1,4 @@
-export const storybookAngularVersion = '5.2.0';
-export const storybookAddonKnobsVersion = '5.2.0';
-export const storybookAddonKnobsTypesVersion = '5.0.3';
+export const storybookAngularVersion = '5.2.5';
+export const storybookAddonKnobsVersion = '5.2.5';
 export const babelCoreVersion = '7.5.4';
 export const babelLoaderVersion = '8.0.6';
